[수정] payment -> price로 변경
